why is the fs error happening????
