+++ utils/data/photos.js
@@ -1,4 +1,4 @@
-import { buildAreaName, determineArea } from '../builders'
+import { buildAreaName } from '../builders'
 import { fmt, getDatabaseQueryConfig, notion } from '../notion'
 
 const getPhotosConfig = (nextCursor = null) =>
@@ -6,6 +6,25 @@ const getPhotosConfig = (nextCursor = null) =>
 const photosSort = [{ property: 'taken_on', direction: 'descending' }]
 
 const formatPhotos = (response) => {
+  const determineArea = (area, area_fallback = null) => {
+    let areaFormatted = fmt(area)
+    let fallbackArea = fmt(area_fallback)
+    if (areaFormatted && areaFormatted != '') {
+      return {
+        region: areaFormatted.split(', ')[0],
+        state: areaFormatted.split(', ')[1],
+      }
+    } else if (fallbackArea && fallbackArea != '') {
+      return {
+        region: fallbackArea.split(', ')[0],
+        state: fallbackArea.split(', ')[1],
+      }
+    }
+    return {
+      region: null,
+      state: null,
+    }
+  }
   return response.map((result) => {
     const {
       id,
